requirements.txt
selenium
selenium-wire
undetected-chromedriver
requests
python-dotenv
